feat: improve context-menu component
- remove UserMenu component
